put everything together to the 987-line behemoth of essentials.py
